+++ database/seeds/DatabaseSeeder.php
@@ -140,6 +140,12 @@ class DatabaseSeeder extends Seeder
             'visual_reference_points',
             'airfield_visual_reference_point',
         ],
+        FirExitPointSeeder::class => [
+            'fir_exit_points',
+        ],
+        IntentionCodeSeeder::class => [
+            'intention_codes',
+        ],
     ];
 
     const OTHER_TABLES_TO_TRUNCATE = [
@@ -147,7 +153,6 @@ class DatabaseSeeder extends Seeder
         'missed_approach_notifications',
         'network_controller_positions',
         'controller_position_alternative_callsigns',
-        'intention_codes',
     ];