+++ lang/en/dependencies/table.php
@@ -1,0 +1,15 @@
+<?php
+
+return [
+    'columns' => [
+        'key' => 'Dependency Key',
+        'local_file' => 'Plugin Local File',
+        'updated_at' => 'Last Updated',
+    ],
+    'actions' => [
+        'download' => [
+            'label' => 'Download',
+            'tooltip' => 'Download Depdendency File',
+        ],
+    ],
+];