+++ lang/en/fir_exit_points/table.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'columns' => [
+        'exit_point' => 'Exit Point',
+        'internal' => 'Internal Exit',
+        'exit_direction_start' => 'Exit Heading Start',
+        'exit_direction_end' => 'Exit Heading End',
+    ],
+];