+++ lang/en/intention/table.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'columns' => [
+        'priority' => 'Priority',
+        'description' => 'Description',
+        'code' => 'Code',
+    ],
+];